+++ src/lightkurve/periodogram.py
@@ -17,13 +17,8 @@ from astropy.units import cds
 from astropy.convolution import convolve, Box1DKernel
 from astropy.time import Time
 
-# LombScargle was moved from astropy.stats to astropy.timeseries in AstroPy v3.2
-try:
-    from astropy.timeseries import LombScargle
-    from astropy.timeseries import implementations  # for .main._is_regular
-except ImportError:
-    from astropy.stats import LombScargle
-    from astropy.stats.lombscargle import implementations
+from astropy.timeseries import LombScargle
+from astropy.timeseries.periodograms.lombscargle import implementations  # for .main._is_regular
 
 
 from . import MPLSTYLE