+++ CHANGES.rst
@@ -1,10 +1,25 @@
-2.4.0 (unreleased)
+2.4.1dev (unreleased)
+=====================
+
+- Made `LombScarglePeriodogram` compatible with Astropy v5.3 [#1342]
+- Updated the TPF plotting function to work correctly with WCS plotting [#1298]
+- Added the ability to open light curves from the TGLC High Level Science Product
+
+2.4.0 (2023-02-14)
 ==================
+
 
 - Added the ability to configure the default cache directory used by
   ``SearchResult.download()`` / ``SearchResult.download_all()``. [#1214]
+
 - Moved the default cache directory from ``$HOME/.lightkurve-cache``
   to ``$HOME/.lightkurve/cache``. [#1214]
+
+- Fixed interact features, e.g. ``tpf.interact()``, to work with bokeh v3.x [#1262]
+
+- Fixed ``SearchResult`` HTML display for links to TESS GI cycles 3/4 proposals. [#1260]
+
+- Fixed references to np.int and np.float due to changes in numpy v1.24. [#1279]
 
 
 2.3.0 (2022-07-07)